Relocate deprecated API shims so the ESM engine tree-shakes
deprecated.js was the one module listed in package.json sideEffects, and
its module-scope prototype patches statically imported AppBase,
StandardMaterial, ForwardRenderer, RigidBodyComponentSystem and more. As
a result, any import from the ESM package retained most of the engine:
a Vec3-only app bundled to 882 KB (229 KB gzip).

Each shim now lives on the class it patches as a regular member with
@ignore/@deprecated JSDoc (matching Entity#getGuid), so it ships only
when the class does and behaves identically. Shims with runtime-computed
names (StandardMaterial vertex-color/tint aliases, StandardMaterialOptions
litOptions forwarding) remain module-scope helper loops, and the
ForwardRenderer#renderComposition patch stays in app-base.js because it
needs getApplication and scene code must not import from framework.
deprecated.js keeps only side-effect-free aliases and helper functions,
and package.json now declares sideEffects: false.

Type declarations: the anisotropy entry is removed from the
rollup-types-fixup synthesis list since the real accessor now provides
the typing (this also fixes its doc text, which was mis-sliced from
anisotropyIntensity). Three StandardMaterial setters assign their target
properties via Object.assign because tsc declaration emit synthesizes
duplicate members from literal this.x assignments to fixup-managed props.

A canary test bundles 'import { Vec3 }' against the built ESM tree and
fails if it exceeds 10 KB, so a future module-scope side effect cannot
silently regress tree-shaking.

Measured (esbuild, minified): Vec3-only app 882 KB -> 3.2 KB
(229 KB -> 1.1 KB gzip); minimal engine-only cube app 1129 KB -> 1056 KB
(297 KB -> 279 KB gzip); full-namespace bundle unchanged.
