+++ package.json
@@ -65,9 +65,7 @@
     "./build/*": "./build/*",
     "./scripts/*": "./scripts/*"
   },
-  "sideEffects": [
-    "./build/playcanvas/src/deprecated/deprecated.js"
-  ],
+  "sideEffects": false,
   "type": "module",
   "bugs": {
     "url": "https://github.com/playcanvas/engine/issues"
