+++ utils/plugins/rollup-types-fixup.mjs
@@ -10,7 +10,6 @@ const TYPES_PATH = './build/playcanvas/src';
 const STANDARD_MAT_PROPS = [
     ['alphaFade', 'boolean'],
     ['ambient', 'Color'],
-    ['anisotropy', 'number'],
     ['anisotropyIntensity', 'number'],
     ['anisotropyRotation', 'number'],
     ['anisotropyMap', 'Texture|null'],
